파이어베이스 Type 이슈 수정

+++ app/src/main/java/org/techtown/kormate/UI/Activity/BoardPostActivity.kt
@@ -125,7 +125,7 @@ class BoardPostActivity : AppCompatActivity() {
     }
 
     private fun uploadPost(post : String ,picUri : MutableList<String> , progressDialog : ProgressDialog) {
-        boardPostViewModel.uploadPost(postsRef, BoardDetail( postId.toString(), userId, userNickName, userProfileImg
+        boardPostViewModel.uploadPost(postsRef, BoardDetail( postId.toString(), userId.toLong(), userNickName, userProfileImg
             , post, picUri, CurrentDateTime.getPostTime() ))
         progressDialog.dismiss()
     }
